+++ rust/net/infra/src/noise/stream.rs
@@ -15,6 +15,7 @@ use snow::TransportState;
 use tokio::io::{AsyncRead, AsyncWrite, ReadBuf};
 
 use crate::noise::{FrameType, Transport};
+use crate::Connection;
 
 /// Stream abstraction that encrypts/decrypts with [Noise].
 ///
@@ -62,6 +63,12 @@ impl<S> NoiseStream<S> {
             read: Read::default(),
             write: Write::default(),
         }
+    }
+}
+
+impl<S: Connection> Connection for NoiseStream<S> {
+    fn transport_info(&self) -> crate::TransportInfo {
+        self.inner.transport_info()
     }
 }